fix: correct division order in CU/SU calculations to avoid integer truncation

+++ grid-client/calculator/calculate.go
@@ -117,20 +117,21 @@ func (c *Calculator) CalculateDiscount(cost float64) (dedicatedPrice, sharedPric
 }
 
 func calculateSU(hru, sru int64) float64 {
-	return float64(hru/1200 + sru/200)
+	return float64(hru)/1200 + float64(sru)/200
 }
 
 func calculateCU(cru, mru int64) float64 {
-	MruUsed1 := float64(mru / 4)
-	CruUsed1 := float64(cru / 2)
+
+	MruUsed1 := float64(mru) / 4
+	CruUsed1 := float64(cru) / 2
 	cu1 := math.Max(MruUsed1, CruUsed1)
 
-	MruUsed2 := float64(mru / 8)
+	MruUsed2 := float64(mru) / 8
 	CruUsed2 := float64(cru)
 	cu2 := math.Max(MruUsed2, CruUsed2)
 
-	MruUsed3 := float64(mru / 2)
-	CruUsed3 := float64(cru / 4)
+	MruUsed3 := float64(mru) / 2
+	CruUsed3 := float64(cru) / 4
 	cu3 := math.Max(MruUsed3, CruUsed3)
 
 	cu := math.Min(cu1, cu2)
